Add disable_conservative_redecl flag
Summary:
When performing an incremental typecheck, we attempt to compare the declarations of definitions that have changed against their previous declaration. This allows us to use fine-grained class member dependency tracking to redeclare and recheck smaller sets of files. However, after an init from a saved state, we do not have old versions of any declarations. When collecting dependencies to redeclare and recheck, we conservatively choose to redeclare and recheck every file that ever used the changed class in ANY way.

This seems *too* conservative. We can't work out any reason why *redeclaring* the entire set of files where the class or its members were referenced (this is what `get_ideps` gives us for a `Dep.Class`) is necessary. If some file has a typing dependency on a changed class (for example, because it invokes a method on an instance of that class), we shouldn't need to *redeclare* it when that class changes; just *recheck* it. Of course, we do need to redeclare classes which *extend* the changed class, but we shouldn't need to redeclare anything else.

Reviewed By: dabek

Differential Revision: D9927186

fbshipit-source-id: eb4efa9e42ab81421e45548f7f6322ce3915dd1f