Delete eval support from preg_replace.
Summary:
There was a sneaky way to do an eval: preg_replace with a pattern including the /e switch.
This was deprecated in PHP 5.5 (with preg_replace_callback (which takes a callable) as the replacement). It was removed in PHP 7.0.
We had this disabled in repo mode, and had it produce deprecation warnings in sandbox mode.
This sneaky way of doing an eval is problematic for Rx enforcement (we want to allow preg_replace in Rx code).
Delete the implementation and have it always throw.

Reviewed By: markw65

Differential Revision: D9985802

fbshipit-source-id: cbe3590d6df63df17f6738b54c26f717b0b87ea9