Emit error for class properties with no visibility modifier
Summary:
Currently, the FFP accepts class properties with no visibility modifier, whereas the legacy Hack parser and the HHVM parser emit errors about unexpected tokens.

This change adds a check to the FFP errors pass which verifies that the modifiers list for a class property is non-empty.

Note that the `var` keyword is considered to be a visibility modifier by the FFP, so this is also the correct behavior in PHP mode, where we accept `var` as a synonym for `public`.

Reviewed By: jamesjwu

Differential Revision: D9695693

fbshipit-source-id: ae45bdef8733949ab83621e6c891ef131e1c5b6b