ConvPtrToLval shouldn't be a passthrough instruction
Summary:
ConvPtrToLval is marked as a passthrough instruction. This is problematic
because ConvPtrToLval produces a value with a type (Lval) incompatible with its
source (Ptr), and also because it may not necessarily be a simple copy of its
source to dest (in wide lval mode). This is unlike all other passthrough
instructions.

This causes issues because things like load-elim and frame-state will sometimes
try to find the least common ancestor of two SSATmps, which peers through
passthrough values. This means we can change an instruction taking a Lval to
take a Ptr instead, which can cause assertion failures (and crashes in wide lval
mode).

So, mark ConvPtrToLval as not being passthrough. Unfortunately this causes a
performance regression because things like memory effects uses canonical() to
find the origin of pointers. Fix this by treating ConvPtrToLval as a special
case in canonical() and peering through it like a passthrough instruction. Add a
comment to canonical() that its not safe to replace its input with its output
(we currently don't do this).

If this pattern becomes more common, we should consider creating a new notion of
passthrough which doesn't imply a no-op copy of src to dest.

(Note: this ignores all push blocking failures!)

Reviewed By: mofarrell

Differential Revision: D10004257

fbshipit-source-id: 23c617d4516e556fd5c8fd527e06612b14dc7411