Improve error recovery for properties missing visibility modifier
Summary:
The FFP test case property_declaration_missing_visibility_modifier.php.errors.php (added in the previous diff) currently produces low-quality error messages and CSTs for properties missing visibility modifiers (when no attributes are present) because error recovery is not very clever for classish body elements which are missing a leading token which indicates what kind of element they are (such as a visibility modifier, or keywords like `async` or `function`, or an attribute list, which indicates that a property or method declaration follows).

This change improves the case of a missing property visibility modifier by speculatively parsing a property declaration and accepting it if no new errors were added (emitting an error about the missing visibility modifier in a later pass).

Because this speculative parsing is only done in an error case, this should have no impact on the performance of parsing well-formed Hack files.

Reviewed By: jamesjwu

Differential Revision: D9695712

fbshipit-source-id: de17b49a06403906c08a33a1bc14b69995c2c870