Join property declarations with attributes on one line when possible
Summary:
The introduction of the `__LateInit` attribute has raised concerns about the formatting of property declarations. Since hackfmt currently always breaks after attribute lists, adding an attribute to each property declaration in a group of several property declarations which previously had none makes the group much sparser and harder to read.

This change joins the declaration with the attribute list when it fits.

Reviewed By: pittsw

Differential Revision: D10014673

fbshipit-source-id: 4a7e0677e9a152e2ab59a297d2d09fb55f2a401b