Truthiness test check
Summary:
Testing the truthiness of an object type by using it in a condition expression or boolean binary operation is always a logic error. Objects which do not extend `SimpleXMLElement` or implement `Container` are always truthy, so there is never a reason to test their truthiness.

This Tast_check bans truthiness tests on types which are known to always be truthy, including `resource`s, functions, tuples, and object types which do not extend `SimpleXMLElement` or implement `Container`.

Reviewed By: CatherineGasnier

Differential Revision: D8786329

fbshipit-source-id: 642cca56a22ea0d4d6c7663c097c0ec5940a2ded