Adjust time taken logging for CHECK_MERGEBASE events
Summary: The goal of this logging is to see how bad are we getting blocked on calls to Mercurial. But since we start the queries in a different place in non-blocking way, the measurement of entire query time is not this information. We want to measure how long were we blocked on `Future.get` instead.

Differential Revision: D10022462

fbshipit-source-id: c75720591c54a1a866f3854da495739623aadd1e