utility to remap memory with huge pages that doesn't fail when mmap() fails
Summary:
Upon mmap() failure, the specified memory range may be unmapped.  This is problematic when we try to replace existing mappings with hugetlb pages. So we try to tolerate it by checking whether the page is mapped, and back it by normal pages if hugetlb page isn't available.

This diff also simplifies the API and remove the flexibility we don't need.  After the change, we always use MAP_READ | MAP_WRITE when grabbing huge pages (and we can always do mprotect() later if needed).
Now, mmap_2m() is used to grab a huge page from a specified node, and remap_2m() is used to replace a previous mapping with huge pages.  remap_2m() won't unmap the specified address range even when mmap() syscall fails.

Reviewed By: ottoni

Differential Revision: D9953432

fbshipit-source-id: 89bd00061ee314d250831faa0ac1828a72e14d97