Use CompactVector for bytecodes
Summary:
CompactVector was originally written to reduce the size of some of
hhbbc data structures, but at the time didn't support a lot of the
operations we did on std::vector<Bytecode>, so I left it as a
vector. Since then, we've added quite a few more operations, so all we
needed were a couple more forms of insert.

Reviewed By: ricklavoie

Differential Revision: D10023600

fbshipit-source-id: 7c86ca949258621759edb5a16c6bfa604c8dcf69