hh_client passes '--from' to hh_server
Summary:
(1) This diff changes ClientStart.start_server so it now passes the --from argument on to hh_server

(2) It changes hh_server parsing so it parses the --from argument and saves it in server global options

(3) It changes logging so that INIT_MON_START logs to HackEventLogger the argv and the proc-stack-if-from-unknown. INIT_MON_START is the very first event that's logged when we invoke hh_server at the command-line, so it seemed the right place. (Previously, the proc-stack was being logged in server init, and argv wasn't being logged).

Note: the proc-callstacks logged by the server aren't informative. That's because when `hh` launches `hh_server` it does so with the `-d` flag, "should_deatch", which means that hh_server starts then spawns the monitor then terminates immediately. And I suppose monitor process doesn't have a PPID. That's a shame. We can address it in a future diff if ever it becomes important.

Reviewed By: dabek

Differential Revision: D10005386

fbshipit-source-id: 6b153f3e936bb9a804f8df8a369e4581fe223abd