Fix coroutine smart constructor
Summary: Currently, the coroutine smart constructor does not correctly identify the __PPL attribute when there are multiple attributes. In `make_attribute_specification`, it parses the entire text between the angle brackets. Instead, we want to parse each item in the list separately. To do this, I added a helper function into generate_full_fidelity.ml and `full_fidelity_syntax.ml` that folds over a `SyntaxList` and applies the function to each `ListItem`'s item in the list.

Reviewed By: arun11111

Differential Revision: D9890039

fbshipit-source-id: 8c5544c737c9900bd2cb43450be86da0aefb06c4