Enforce that redeclared properties consistently declare __LateInit
Summary:
In the runtime, late init behavior is inherited when properties are redeclared in subclasses. This enforces that users consistently declare redeclared properties to be `__LateInit`.

This will solve the bug in which a redeclared property has a default value, which is a fatal error when the class is being loaded.

Reviewed By: vassilmladenov

Differential Revision: D10055183

fbshipit-source-id: c1aad0589f70a6ec5449d08f8c9b029b1f642cdf