Rewrite small fun with fewer @@
Summary:
I think this function was too complicated before. Too many @@. And there's no need to use Core_result.join if you don't introduce the Ok in the first place.

(I've been trying to refactor around it, but was stuck trying to understand this function, which is why I'm simplifying it).

Reviewed By: arxanas

Differential Revision: D10055604

fbshipit-source-id: 04406f08de3378a056e2d46a9d7b38e661f0e5e5