serverInit refactoring 6
Summary:
In this refactoring, I observe that "invoke_approach" returned
```
unit -> (unit -> (loaded_info, exn) result, 'a) result
```

The final `'a` and `result` was because all codepaths always returned `Ok ...`. Therefore, (1) there was never any `Error` returned from the function and that's why it inferred `'a` for the error type, and (2) it's entirely pointless to return the outer result.

So I made it only return its inner result directly, not wrapped up in an outer result. This also involved changes to with_loader_timeout.

I also added comments about why inoke_approach returns a double-lambda.

Reviewed By: dabek

Differential Revision: D10076290

fbshipit-source-id: 46152a621fa6ca94548159cfad6765d9989e0efd