Avoid spinning when Unit-loading is contended
Summary:
When there is high contention for loading new Units out of the repo,
using a nonconst tbb::chm accessor can cause many CPUs to start spinning
waiting for lock acquisition.  Use a futex and a const_accessor instead.

This locking model is a little funky, but it lets us use SmallLock for
the per-entry synchronization, while leaving the full-table concurrency
to tbb::chm.

Reviewed By: markw65

Differential Revision: D9996598

fbshipit-source-id: fe33e39c551bfe7554ae7ea8c970ec5b2b5fab19