Fix type-checking of switch statements with the default case not listed last
Summary: We currently don't check switch statement cases following the default case.  The default case may be empty and fall through to next non-default cases.  This is arguably a dubious programming pattern but from the perspective of the type system I think the default case should be handled similarly to non-default cases.  We could improve our detection of dead code when the default case is terminal, but we should not allow junk in the non-default cases following a terminal default case either, in my opinion.

Reviewed By: andrewjkennedy

Differential Revision: D10123943

fbshipit-source-id: 71183b38480187b1bd46084f86235c6abe3f0671