Add experimental flag for unresolved-as-union
Summary:
Small step in the journey to type inference based on constraint solving. Introduce new .hhconfig and command-line
flag `unresolved-as-union` that will eventually retire the use of `Tunresolved` except for sound unions. For now,
inhibit the initial binding of a `Tvar` to an empty unresolved.

To see its effect, run the tests with an `HH_FLAGS` file containing `--unresolved-as-union`. Around 50 tests fail: those that make use of "growing" of unresolveds under a type variable.

Reviewed By: arun11111

Differential Revision: D10127332

fbshipit-source-id: 93b54c424778c6fbf4069f3b08402e2411bce671