Add support for FCallArgs async eager offset to HHBC binary representation
Summary:
Extend the HHBC binary representation of FCallArgs with async eager offset and
implement parsing and emitting of it.

HHBBC ignores the async eager offset on purpose, as it would otherwise split
FCall and Await into separate blocks, disabling the FCallAwait optimization.

The next diff that replaces the FCallAwait optimization will undo this.

Differential Revision: D9646712

fbshipit-source-id: ab7d32729cc0ebf7aeb54ce009f83007fee4696d