Give Shapes::toDict() an ad hoc typing rule similar to Shapes::toArray()
Summary: This fixes a discrepancy between how `Shapes::toArray()` and `Shapes::toDict()` are handled by the type-checker.  The former has a special ad hoc typing rule: essentially, union the key types, union the value types, and declare the return type to be an array built out of the unioned types.  `Shapes::toDict()` is just typed according to its type-signature declared in `Shapes.hhi`, so that the return type is always `dict<arraykey, mixed>`, even when we know e.g. that the input shape's keys are all strings.

Reviewed By: kmeht

Differential Revision: D10145888

fbshipit-source-id: d33b9fd9319c5fba4a53318e1bfff571f3b4ef9e