Move the rvalue check into a TAST check
Summary:
We check in many places if an expression has type `void`, `noreturn`, or a union containing any of these two types.  The idea is that an expression of type `void` or `noreturn` cannot be used in a context where a value is required because the types `void` and `noreturn` can only arise as return type hints of functions that don't return values.  This check seems to be a good candidate for extracting into a TAST check, which is what this diff is trying to achieve.

The approach is roughly as follows: when traversing the AST we maintain a boolean flag indicating whether an expression of type `void` or `noreturn` is allowed in a given context.  For example, an expression used as a standalone statement can have types `void` and `noreturn`.  Similarly, and expression in the initialization and update parts of a `for` loop can have type `void` and `noreturn`.  If a ternary if-expression occurs as a standalone statement, i.e., in a context allowing `void` and `noreturn`, then `void` and `noreturn` should also be allowed in the branches.

I'm not sure I've covered all the cases and have done this correctly, hence RFC.

Reviewed By: andrewjkennedy, fahndrich

Differential Revision: D10107312

fbshipit-source-id: eade23a66abae6a941565e0f447db73900251918