Do not waste time constructing types for collections in a mixed or untyped context
Summary:
Previously we would construct union types for expression such as `vec[e1, ..., en]` even if they are used in a context that is expected to be of type `mixed` or "any". Instead, we should expect an element type of `mixed` or any for the elements of the collection. We get significant performance wins on large collection expressions.

A superior approach is to use the subtype constraint solver to solve a constraint `vec<v> <: expected_type` with a fresh variable `v`. Either we find such a `v`, or fail the constraint (so it's an error even without looking at the elements), or the constraint is valid, in which case we just choose `mixed` or any for `v`. We will do this in a future diff.

Reviewed By: manzyuk

Differential Revision: D10192061

fbshipit-source-id: 2455903cc6c2195305aceb983b4c4043c4ab40a3