Make format of hh_show less brittle
Summary:
We use `hh_show` in a lot of our tests. Ideally, we shouldn't do this, except as a cheap way of testing the "hover type" feature. (We should only care whether or not something type-checks, not the precise types that Hack inferred.) But at least let's make it less brittle by not exposing internals of type variables in the
pretty-printed format. (If you want to see internals such as type variables, you can use `hh_show_env`.)

Also change display of "exact" types. `C as C` is confusing, so we now display this simply as `C`, but show it as `exact C`.

Reviewed By: manzyuk

Differential Revision: D10142259

fbshipit-source-id: 6dde169ffe03b329597c43d2d207801b1eff58aa