Operators <= and >= must not evaluate array/object elements twice.
Summary:
We had a subtle bug with the <= and >= operators on arrays and objects: in the interpreter (for arrays; in both the interpreter and jit for objects) we would separately evaluate < and = (or > and =) and logical-or the results together. This could lead to invoking the __toString method twice on a single array element/object property for a single comparison https://3v4l.org/VKFgc
Fixing this also revealed that we were expecting the wrong thing in one edge case in all_type_comparison_test.php so adjust the expectation https://3v4l.org/3O8LUp
Add a new test specifically for this bug.

Reviewed By: markw65

Differential Revision: D10215888

fbshipit-source-id: 3c66aee84910eb27af40a3b3797b9a5e5c7e7dc5