Remove 86reifiedinit from index.methods as well
Summary:
Turns out there is another place that it needs to be removed from after D10110000. This was causing a crash in prod builds after the next stack of diffs in `init_return_type` because we were falling off of the vector.

Also move this function before we define func families.

Reviewed By: markw65

Differential Revision: D10175992

fbshipit-source-id: 34a1d1674836eace8a0315b86e65c5d6429fd3ba