Add an option to raise notices when implicitly calling __toString.
Summary:
Add the Eval.NoticeOnImplicitInvokeToString runtime option, and raise notices when ObjectData::invokeToString implicitly calls `__toString` on an object.
Add a unit test that covers the most common code paths that will trigger this, and have it also verify that explicitly calling `__toString` on an object will not raise the notice.

Reviewed By: ricklavoie

Differential Revision: D10215889

fbshipit-source-id: 58ca1dbaa274d0eb455af39f8bf3bd62e03248c8