Fix some issues with address sizes
Summary:
Each compilation unit has an "address size" field which identifies the
size of addresses on the target architecture. Most addresses are
recorded using that size. DW_FORM_ref_addr is unusual, in that in
dwarf-2 and earlier, it was recorded using the address size; but in
dwarf-3 and later its 32 bits in 32-bit cu's and 64 bits in 64-bit
cu's.

Reviewed By: ricklavoie

Differential Revision: D10234839

fbshipit-source-id: 95a384806cb0144e0d9ad66ac4a306383755f873