don't truncate init_proc_stack so much
Summary:
I was reading a load of init_proc_stacks which were truncated before the most important part, namely the name of the script that invoked it. We're truncating at the first 300 characters.

This diff changes it to the first 1000 characters.

We've been logging about 750k init_proc_stack events per day for a week now. (Prior to an intentional diff it was 150k/day). Total storage of this table at the moment is 2GB, so it will end up at 6gb once we have a full month of 750k events/day.

What will be the effect of this diff? We're currently not truncating every row of each init_proc_stack. From a random sample, it looked like we're truncating maybe half of the rows. So I'd expect total storage for this column to grow to 9gb as a result of this diff.

Currently we're using 283gb/593gb. So we have headroom.

Reviewed By: 2BitSalute, dabek

Differential Revision: D10249246

fbshipit-source-id: 1171eaf01d7be31cef6708ad8ce54a2e69232fac