Fix calling convention in Map collection map/filter/take functions.
Summary:
The Map/ImmMap collection map/mapWithKey/filter/filterWithKey/takeWhile functions that are implemented in hhas use the old, FIsParamByRef calling convention to invoke the callback, but they are marked as `__Rx`, and `__Rx` functions must not use VGetL. We can't change these to use FThrowOnRefMismatch because non-Rx code may legally call these with a callback that has by-ref parameters.
Jan convinced me that it is OK for these functions to always pass cells to the callback, regardless of if it wants params by ref or by value (i.e. don't check the param reffiness using either FIsParamByRef or FThrowOnRefMismatch). This diff implements this "pass ignoring desired reffiness" approach.
This produces a behavior change, but the old behavior was not useful: since you can't bind collection elements, if you took params by ref then all that the callback's params would be bound to is the locals in the map/filter/take function, so the only way to even observe the binding would be to do something silly like this: P60160919

Reviewed By: jano

Differential Revision: D10247964

fbshipit-source-id: 2c5b90e90d9f7988ae4210d47d1456541488a412