Reverse mapping direction in Typing_deps.dump_debug_deps
Summary: This debug function prints dependency edges in the reverse direction of what is actually stored in the dependency table. jezng wrote in a comment that he found this direction less confusing, but I (and, IIRC, ljw1004) don't! Since when we read the dependency table, we are generally answering the question "given that X has changed, what must be rechecked?", I think it's strange to print the table in a way that makes it harder to answer that question (since it is currently sorted by the dep which must be rechecked, rather than the dep which has changed). It also makes more sense to me to have the arrows pointing in the direction of information flow.

Reviewed By: dabek

Differential Revision: D10247118

fbshipit-source-id: b2b44765d0eb955d15a9803aa864fe67c4abf9a4