Add integration test for disable_conservative_redecl
Summary:
In this test, we expect these dependencies (among others) to be stored in the saved state (where the arrow is read, "is used by", or "is a dependency of"):

```
  Class Foo -> Class Bar
  Class Bar -> Class Qux
```

So when Foo is changed (and we have no old decl to compare against because we have loaded from a saved state), and we have disabled the conservative redecl behavior, we only redeclare classes which extend Foo (and there are none). We must recheck both Foo and Bar, but we don't need to recheck Qux.

Without disable_conservative_redecl, we would redeclare all of Foo's typing dependencies, including Bar. Since we redeclared Bar (with no old decl to compare against), we would then also redeclare and recheck Qux.

This test verifies that we do not recheck Qux when the disable_conservative_redecl flag is on.

Reviewed By: dabek

Differential Revision: D10243566

fbshipit-source-id: 08d6afacf538aed13b1bcdbd4d2a041caa0515d7