Add DCE/partial-DCE support for InitCtx in inlined closures
Summary: I noticed that when we inline a closure we often fail to elide the frame because of an InitCtx. We can avoid this for cases where the input is a constant (either a TCls or a TNullptr) in pdce-inlining, and we should be able to always drop it during normal DCE frame elision (where the frame will never need to be rematerialized).

Reviewed By: jano

Differential Revision: D10052203

fbshipit-source-id: d1ee1a0faededf7129b6cb47b01885900e37dfe2