Do not use parent frame in syncVMRegs
Summary:
The function syncVMRegsWork, responsible for walking the stack to find
the last sync point, had logic to compensate for a PowerPC issue. Since
the PowerPC frame pointer points to the bottom of the stack, retrieving
the frame pointer at the caller and then using it later in a callee that
was optimized by a tail call optimization would trigger a bug, since this
fp is not valid anymore. The solution at the time was to always retrieve
the frame pointer of the parent frame, as that would be valid. Remove
this logic since this is not the case anymore, as this caller-callee
pair cannot be optimized by tail call in the current code. This fixes a
bug where an aggressive inlining strategy would inline all C++ functions
called from TC, in which case retrieving the parent frame pointer would
not get a frame pointer from a C++ function, necessary for the
algorithm used to sync VM regs.

Reviewed By: markw65

Differential Revision: D10251402

fbshipit-source-id: 1492781c8622e2fc98acd487194092b70cf1b263