ParserOptions.make
Summary: We are beginning to add more parser options as we remove PHP features from Hack. Instead of having a separate `with_*` function for each option, this just makes the record upfront. In the typechecker, we'd use `GlobalOptions.make` to make both the parser options and typechecker options at once, but in the compiler, all we need is `ParserOptions.make`.

Reviewed By: jamesjwu

Differential Revision: D10235606

fbshipit-source-id: 2a1acf2e762a2c68168f6a2a7557c7ee8df4c45b