Represents reified classes with unique names
Summary:
This diff adds functionality to represents reified classes in the VM with unique names by giving the instantiated class a unique entry on the NamedEntity table mangled with its reification, i.e. class C<T1,T2> when reified with T1 = `int` and T2 = `(string,int)` becomes `C$$<int,(string,int)>` where this entry still points to original class definition.

It also adds the functionality to store this reification on a global reified generics table mapping the string `<int,(string,int)>` to a `ArrayData*` which contains a static varray/vec of runtime representation of these reified generics.

This diff does not make use of this mapping but the following diffs will make use of this functionality.

Followups:

1) Use the same mechanism for functions.
2) Use this mechanism to access the reification of classes in the VM
3) Ditto for functions

I'll be adding these followups on the following diffs.
This diff also has an additional bytecode instruction. `RecordReifiedGeneric` which is also not used in this particular diff. The implementation is subsumed by `ReifiedName` instruction hence I chose to add it to this diff. In the next diff, this instruction will be used in `86reifiedinit` method.

Reviewed By: paulbiss

Differential Revision: D10153546

fbshipit-source-id: 33c6c8b196d36594d58c8d7bbf5577acf61fb8d7