Represent reified functions/methods with unique names
Summary:
Similar to what we did in D10153546 for classes, we repeat the same unique names for functions and methods in this diff. This is done by reusing the machinery built in D10153546.

The goal of this diff is to enable extraction of reified generics from the name of the function for the next diff on the stack.

Reviewed By: paulbiss

Differential Revision: D10242446

fbshipit-source-id: 1ccc4a60df452c90b103f4cfa61465840fafa614