Measure time spent in GC during handling of IDE commands
Summary:
Based on {D5264805}. That diff added this profiling unconditionally to all worker runs, so I'm assuming that it is very cheap (and it looks to be so), and I can also add it unconditionally to all IDE queries.

I needed to add some resetting / removing of hooks that was not needed in worker case because there it's handled by worker being short-lived.

Reviewed By: pittsw

Differential Revision: D10362609

fbshipit-source-id: c62e5b6735bcca880e302c878be7511048706f72