Ban multiple concrete interface constant definitions
Summary:
Currently, overriding interface constants is banned in HHVM, but I want to change that as well. What should *always* be banned is inheriting multiple concrete constants from interfaces without overriding, just like with traits.

Thus we ban it here the same way we banned multiple concrete trait definitions.

Also with this diff, I changed the wording of our error message to say "members" instead of "methods" because we use the same error message for constants, and methods is definitely not correct.

Reviewed By: kmeht

Differential Revision: D10363512

fbshipit-source-id: 30400280591cd0795a7f741e983595782f923bd3