Improve vasm DCE
Summary:
Add special handling for copyargs and phijmp/phidef to the vasm dead code
elimination pass. If these are treated like normal instructions, we'll only
dead-code eliminate them if all of their dests are dead. If we recognize and
handle them as a special case, we can do better. We can propagate the liveness
of each individual dest through to its specific individual source. Moreover, if
that dest is dead, we can remove it from the copy individually. For phis, if all
the dests are removed, we can lower the phijmp/phidef pairs into just a jmp
(which may allow for further simplification).

Also simplify the pass a bit. It was keeping live-in sets for each block, but
this is unnecessary. Whether a particular variable is dead is flow insensitive
(because we're in SSA form), so we can use a single bitset for the entire unit.

Reviewed By: markw65

Differential Revision: D10305856

fbshipit-source-id: cc70c00a83a1176224779aa55a3c9442730717f8