Smarter unification of unresolved types
Summary:
When we unify two unresolved types, say `Tunresolved tyl1` and `Tunresolved tyl2`, we effectively first construct an unresolved type containing both `tyl1` and `tyl2`, and then we eliminate redundant elements: if we have a pair of types `ty1` from `tyl1` and `ty2` from `tyl2` that unify, we only need to keep one of the types in the result.

This diff takes this optimization one step further: if one type is a subtype of the other, we only need to keep the supertype (because the subtype is contained in the supertype and hence in the union).

To perform the subtyping test, we use `is_sub_type_alt`, which only does solution-preserving simplification of subtyping constraints and doesn't perform any unifications.

This optimization allows us to avoid creating unnecessarily large (sometimes even exponentially large) unresolved types.  For example, without this optimization (and with the changes to the array assignment typing from D10258204), in the following case:
```
type s = shape('x' => ?int, 'y' => ?int);

function test(s $s): void {
  $v = Vector {$s};
  $v[0]['x'] = 0;
  $v[0]['y'] = 0;
}
```
the type of `$v` at the end of the function becomes `Vector<(shape('x' => int, 'y' => int) | shape('x' => int, 'y' => ?int) | shape('x' => ?int, 'y' => int) | shape('x' => ?int, 'y' => ?int))>`.  It's instructive to see why:
- After the first line, the type of `$v` is `Vector<(shape('x' => ?int, 'y' => ?int))>`.
- After the assignment `$v[0]['x'] = 0;` we know that the first element of the vector has a more precise type `shape('x' => int, 'y' => ?int)` (the `'x'` field is guaranteed to be set to a non-null value), but we don't know anything about the other elements, so the overall type of elements of the Vector is a union `shape('x' => int, 'y' => ?int) | shape('x' => ?int, 'y' => ?int)`.
- To process the assignment `$v[0]['y'] = 0;`, we have to consider all possible cases:
    - If the element type is `shape('x' => int, 'y' => ?int)`, then the type of the first element after the assignment becomes `shape('x' => int, 'y' => int)` and the types of the rest of the elements remain the same, so the element type in this case becomes `shape('x' => int, 'y' => int) | shape('x' => int, 'y' => ?int)`.
    - Similarly, if the element type is `shape('x' => ?int, 'y' => ?int)`, then upon the assignment it becomes `shape('x' => ?int, 'y' => int) | shape('x' => ?int, 'y' => ?int)`.

It's easy to see how this becomes exponential in the number of fields.  Note, however, that the created unions are all equivalent to `shape('x' => ?int, 'y' => ?int)`, but the old algorithm can't figure this out because it tries to unify say `shape('x' => int, 'y' => ?int)` with `shape('x' => ?int, 'y' => ?int)` and these types don't unify.

Note also that we already perform the same optimization when computing unions.

Reviewed By: CatherineGasnier

Differential Revision: D10377988

fbshipit-source-id: ec3f0ecdd8ae31f797bf8d9cefb99419dce3add5